+++ reability.go
@@ -2,7 +2,7 @@ package metatext
 
 import "fmt"
 
-func (metaData *MetaData) CalculateReadabilityIndex(index int) (uint, error) {
+func (metaData *MetaData) CalculateReadabilityIndex(index int) (float32, error) {
 	switch index {
 	case 0:
 		return metaData.CalculateLix(), nil
@@ -16,16 +16,16 @@ func (metaData *MetaData) CalculateReadabilityIndex(index int) (uint, error) {
 }
 
 // 0-55+
-func (metaData *MetaData) CalculateLix() uint {
+func (metaData *MetaData) CalculateLix() float32 {
 	return metaData.WordCount/metaData.PeriodCount + (metaData.LongWordCount*100)/metaData.WordCount
 }
 
 // 100-0
-func (metaData *MetaData) CalculateFleschReading() uint {
-	return uint(206.835 - 1.015*float64((metaData.WordCount/metaData.PeriodCount)) - 84.6*float64((metaData.SyllableCount/metaData.WordCount)))
+func (metaData *MetaData) CalculateFleschReading() float32 {
+	return 206.835 - 1.015*(metaData.WordCount/metaData.PeriodCount) - 84.6*(metaData.SyllableCount/metaData.WordCount)
 }
 
 // 0-18
-func (metaData *MetaData) CalculateFleschGrade() uint {
-	return uint(0.39*float64((metaData.WordCount/metaData.PeriodCount)) + 11.8*float64((metaData.SyllableCount/metaData.WordCount)) - 15.59)
+func (metaData *MetaData) CalculateFleschGrade() float32 {
+	return 0.39*(metaData.WordCount/metaData.PeriodCount) + 11.8*(metaData.SyllableCount/metaData.WordCount) - 15.59
 }
